Fixing material design references
The main issue is the project was using .NET Core 3.0. This version reached end of life back in March 2020. https://devblogs.microsoft.com/dotnet/net-core-3-0-end-of-life/

I also added in the basics for font size, and roboto font in the MainWindow.

Also inside of the csproj it appears that MovementChartView may have been inadvertently added with default settings. I removed these as they should not be needed.